Reorginazation of the files and some type annotations
