+++ Exercises/BoWichtendahl/exercise_07/exercises_07.py
