+++ Exercises/BoWichtendahl/exercise_07/int_computer.py
